Make experiment code robust to unenrollment.
When users unenroll from experiments, their telemetry data no longer
mentions the experiment and so with the old approach these users
appear to drop off the face of the earth. Mass unenrollment in one
variation (which is bad!) appears as mass churn (which is worse!).

Accordingly, "make sure to query data for unenrolled people" was bumped
up the priority table.

Metrics now typically include data from unenrolled people; this is usually
The Correct Thing. If it's not right for your analysis then you can
define an awkward-looking metric that zero-fills data for unenrolled
users - you just need to do it explicitly.