Document and tidy experiment query code.
The main practical change is that `.cache()` and `F.broadcast()` are now called on the `enrollments` DataFrame it is returned. If this proves problematic for certain usecases then we can make it optional or remove it, but I figured let's not add an option that we suspect no-one will ever use.

The other outward-facing change is that `get_enrollments()` has become a 'universal' method that returns an `enrollments` DataFrame both for pref flip and for addon experiments. In the future it would also be used for glean experiments. I figured a single entrypoint was more discoverable than separate methods for each method - plus this eliminates a few lines of repeated code.

The `check_consistency` stub method has been removed; I think consistency checks are going to have to live in another module since eventually there may be many of them and they smell different to this module's contents.